Add ability to parse github format tables
Takes a string formated per the github version of markdown. It returns an html table if the first character in the first line is a pipe. Otherwise is returns the string. See https://help.github.com/articles/organizing-information-with-tables/ except that the pipes on either side are not optional. The columns can be aligned right, center or left if colons are inserted on the sides of the hyphens within the header row. At the moment you cannot use a pipe as content within the cell.